Update unit-test of Customize_Snapshot_Manager
